requirements.txt

requests
opencv-python
Pillow
numpy
python-dotenv
llama-index
llama-index-llms-openai
